awsdawdasdawswd1234123asd
awqda

+++ 1_Data_Access_2/7_Hibernate/Parte2_Final/src/main/java/com/hospitalcrud/ui/RestPatient.java
@@ -1,6 +1,9 @@
 package com.hospitalcrud.ui;
 
+import com.hospitalcrud.domain.error.MedicalRecordException;
+import com.hospitalcrud.domain.model.MedRecordUI;
 import com.hospitalcrud.domain.model.PatientUI;
+import com.hospitalcrud.service.MedRecordService;
 import com.hospitalcrud.service.PatientService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -15,9 +18,11 @@ import java.util.List;
 public class RestPatient {
 
     private final PatientService patientService;
+    private final MedRecordService medRecordService;
 
-    public RestPatient(PatientService patientService) {
+    public RestPatient(PatientService patientService, MedRecordService medRecordService) {
         this.patientService = patientService;
+        this.medRecordService = medRecordService;
     }
 
     @GetMapping
@@ -25,6 +30,11 @@ public class RestPatient {
         return patientService.getPatients();
     }
 
+    @GetMapping
+    @RequestMapping("/{patientId}/medRecords")
+    public List<MedRecordUI> getPatientMedRecords(@PathVariable int patientId) {
+        return medRecordService.getMedRecords(patientId);
+    }
 
     @PostMapping
     public int addPatient(@RequestBody PatientUI patientUI) {
@@ -40,6 +50,9 @@ public class RestPatient {
     @RequestMapping("/{patientId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public boolean deletePatient(@PathVariable int patientId, @RequestParam boolean confirm) {
+        if (!confirm && !medRecordService.checkPatientMedRecords(patientId)) {
+            throw new MedicalRecordException("Patient has medical records, cannot delete.");
+        }
         return patientService.delete(patientId, confirm);
     }
 }
